+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldAblazeEnchantment.java
@@ -1,14 +1,13 @@
 package com.insane96mcp.shieldsplus.world.item.enchantment;
 
 import com.insane96mcp.shieldsplus.setup.SPEnchantments;
+import com.insane96mcp.shieldsplus.world.item.SPShieldItem;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 
 public class ShieldAblazeEnchantment extends Enchantment {
@@ -16,7 +15,7 @@ public class ShieldAblazeEnchantment extends Enchantment {
 	public static final int SECONDS_ON_FIRE = 3;
 
 	public ShieldAblazeEnchantment() {
-		super(Rarity.UNCOMMON, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
+		super(Rarity.UNCOMMON, SPShieldItem.SHIELD, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
 
 	public int getMinCost(int p_44598_) {
@@ -29,11 +28,6 @@ public class ShieldAblazeEnchantment extends Enchantment {
 
 	public int getMaxLevel() {
 		return 2;
-	}
-
-	@Override
-	public boolean canApplyAtEnchantingTable(ItemStack stack) {
-		return stack.getItem() instanceof ShieldItem;
 	}
 
 	public static void onBlocked(LivingEntity blockingEntity, DamageSource source) {